+++ src/DotPulsar/SinglePartitionRouter.cs
@@ -46,10 +46,10 @@ namespace DotPulsar
         /// <summary>
         /// Choose a partition in single partition routing mode
         /// </summary>
-        public int ChoosePartition(MessageMetadata? messageMetadata, int numberOfPartitions)
+        public int ChoosePartition(MessageMetadata messageMetadata, int numberOfPartitions)
         {
-            var keyBytes = messageMetadata?.KeyBytes;
-            if (keyBytes is not null)
+            var keyBytes = messageMetadata.KeyBytes;
+            if (keyBytes is not null && keyBytes.Length > 0)
                 return (int) MurmurHash3.Hash32(keyBytes, 0) % numberOfPartitions;
             
             if (_partitionIndex == -1)