+++ CHANGELOG.md
@@ -2,7 +2,32 @@
 
 All notable changes to this project will be documented in this file.
 
-The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [Unreleased]
+
+## Added
+
+- [Tracing](https://github.com/apache/pulsar-dotpulsar/wiki/Tracing) support following the [guidelines](https://github.com/open-telemetry/opentelemetry-specification/blob/main/specification/trace/semantic_conventions/messaging.md) from the [OpenTelemetry](https://opentelemetry.io/) project
+    - Sending a message will create a producer trace and add tracing metadata to the message
+    - The 'Process' extension method for IConsumer\<TMessage\> is no longer experimental and will create a consumer trace
+
+### Changed
+
+- **Breaking**: Sending a message without metadata is now an extension method and therefore no longer part of the ISend\<TMessage\> (and thereby IProducer\<TMessage\>) interface
+- IMessageRouter: ChoosePartition(MessageMetadata? messageMetadata, int numberOfPartitions) -> ChoosePartition(MessageMetadata messageMetadata, int numberOfPartitions)
+
+## [1.1.2] - 2021-07-05
+
+### Fixed
+
+- The partitioned producer didn't pass metadata to the message router
+
+## [1.1.1] - 2021-07-05
+
+### Fixed
+
+- The producer ignored message metadata provided by the user
 
 ## [1.1.0] - 2021-06-29